+++ src/main/java/org/mariadb/jdbc/internal/com/send/parameters/OffsetTimeParameter.java
@@ -106,7 +106,7 @@ public class OffsetTimeParameter implements Cloneable, ParameterHolder {
     pos.write(QUOTE);
   }
 
-  public long getApproximateTextProtocolLength() {
+  public int getApproximateTextProtocolLength() {
     return 15;
   }
